refactor: extract redact.ts shared module; guard non-plain objects; fix config.ts debug leak
Extends #140 with three follow-up improvements from Gemini review:

1. Extract SENSITIVE_KEY_PATTERN and redactSecrets to src/lib/redact.ts so
   config.ts can import the util without a circular dependency
   (index → config → index would result from importing the other direction).
   Both index.ts and config.ts now import from ./lib/redact.ts; redact.ts
   imports nothing from this project.

2. Guard redactSecrets against non-plain objects: Date/RegExp/Map/class
   instances were previously rebuilt as {} on the recursive branch. Now only
   plain objects (proto === Object.prototype or null) are walked; everything
   else is returned as-is.

3. Fix the config.ts pluginDefaults debug-log secret leak: wrap
   logger?.debug?.('Plugin defaults:', pluginDefaults) with redactSecrets()
   so a plugin-level clientSecret isn't emitted at debug level.

Test: update import path to dist/lib/redact.js; add 4 non-plain-object
guard cases (Date, RegExp, Map, mixed plain+Date object).
