+++ src/index.ts
@@ -12,30 +12,8 @@ import { clearOAuthSession } from './lib/handlers.ts';
 import { HookManager } from './lib/hookManager.ts';
 import { DynamicProviderCache, DEFAULT_DYNAMIC_PROVIDER_CACHE_TTL_SECONDS } from './lib/dynamicProviderCache.ts';
 import { registerWellKnownHandlers } from './lib/mcp/wellKnown.ts';
+import { redactSecrets } from './lib/redact.ts';
 import type { Scope, OAuthPluginConfig, ProviderRegistry, OAuthHooks } from './types.ts';
-
-// Config can carry literal secrets (provider clientSecret/client_secret,
-// mcp.signingKeyPem/signing_key_pem, mcp.dynamicClientRegistration.initialAccessToken/
-// initial_access_token, and generic private_key/api_key/passphrase/credential).
-// Redact them before logging the options blob — logs are frequently shipped/retained
-// outside the trust boundary. Deny-list by key-name substring; over-redaction in a
-// log is safe. Patterns use optional [_-]? separators to catch camelCase, snake_case,
-// and kebab-case variants in one pass. Bare `token` and `key` are intentionally
-// excluded — they would redact non-secret values like refreshTokenTtl or kid.
-export const SENSITIVE_KEY_PATTERN =
-	/secret|signing[_-]?key|private[_-]?key|api[_-]?key|initial[_-]?access[_-]?token|password|passphrase|credential/i;
-
-export function redactSecrets(value: unknown): unknown {
-	if (Array.isArray(value)) return value.map(redactSecrets);
-	if (value && typeof value === 'object') {
-		const out: Record<string, unknown> = {};
-		for (const [key, val] of Object.entries(value as Record<string, unknown>)) {
-			out[key] = SENSITIVE_KEY_PATTERN.test(key) ? '[REDACTED]' : redactSecrets(val);
-		}
-		return out;
-	}
-	return value;
-}
 
 // Export HookManager class, OAuthResource class, and types
 export { HookManager } from './lib/hookManager.ts';
